Change brush chart with timeseries chart
